+++ examples/components/datagrid/datagrid.component.jsx
@@ -131,7 +131,6 @@ export default class DatagridView extends React.Component {
           columns={columns}
           disableActionSave={disableActionSave}
           actionBarLeft={actionBar}
-          enableArrowNavigation
           filtering
           inlineEdit
           removing