+++ src/datagrid/datagrid.component.jsx
@@ -273,8 +273,13 @@ export default class DataGrid extends React.PureComponent {
     }
   }
 
-  onCellFocus = (e) => {
-    e.target.select();
+  onCellFocus = (cellType, inputType, rowIndex, columnKey) => (e) => {
+    if (inputType === 'number') {
+      e.target.select();
+    }
+    if (cellType === 'edit' && this.props.cellSelect) {
+      this.props.cellSelectionChange(this.props.grid, Map({ rowIndex, columnKey }));
+    }
   }
 
   getDataIdByRowIndex = rowIndex =>
@@ -453,9 +458,16 @@ export default class DataGrid extends React.PureComponent {
 
   // Tweak to get array function away from render
   handleEditCellRef = (rowIndex, col) => (ref) => {
+    const columnKey = Utils.getColumnKey(col);
     if (this.focusToEditCell && !this.getComponentDisabledState(rowIndex, col, 'edit')) {
       const selectedRowIndex = this.getSelectedItemIndex(this.props.selectedItems.first());
-      if (selectedRowIndex === undefined) {
+      const selectedCell = this.props.selectedCell;
+      if (selectedCell.size > 0) {
+        if (selectedCell.get('rowIndex') === rowIndex && selectedCell.get('columnKey') === columnKey) {
+          ref.focus();
+          this.focusToEditCell = false;
+        }
+      } else if (selectedRowIndex === undefined) {
         this.focusToEditCell = false;
       } else if (selectedRowIndex === rowIndex) {
         ref.focus();
@@ -463,7 +475,6 @@ export default class DataGrid extends React.PureComponent {
       }
     }
     if (this.props.enableArrowNavigation) {
-      const columnKey = Utils.getColumnKey(col);
       this.cellRefs[`${this.props.grid.id}_${columnKey}_${rowIndex}`] = ref;
     }
   }
@@ -603,6 +614,7 @@ export default class DataGrid extends React.PureComponent {
                     value={this.getEditItemValue(rowIndex, col)}
                     onChange={this.onEditCellValueChange(rowIndex, col, editValueParser)}
                     onBlur={this.onEditCellBlur(rowIndex, col)}
+                    onFocus={this.onCellFocus('edit', col.componentType, rowIndex, column.columnKey)}
                     onKeyDown={this.onEditCellKeyDown(col, rowIndex)}
                     inputRef={this.handleEditCellRef(rowIndex, col)}
                     id={`ocDatagridEditInput-${this.props.grid.id}-${column.columnKey}-${rowIndex}`}
@@ -656,7 +668,7 @@ export default class DataGrid extends React.PureComponent {
                     value={this.getEditItemValue(rowIndex, col)}
                     onChange={this.onEditCellValueChange(rowIndex, col, editValueParser)}
                     onBlur={this.onEditCellBlur(rowIndex, col)}
-                    onFocus={this.onCellFocus}
+                    onFocus={this.onCellFocus('edit', col.componentType, rowIndex, column.columnKey)}
                     onKeyDown={this.onEditCellKeyDown(col, rowIndex)}
                     inputRef={this.handleEditCellRef(rowIndex, col)}
                     id={`ocDatagridEditInput-${this.props.grid.id}-${column.columnKey}-${rowIndex}`}
@@ -674,7 +686,7 @@ export default class DataGrid extends React.PureComponent {
                     value={this.getCreateItemValue(rowIndex, col)}
                     onChange={this.onCreateCellValueChange(rowIndex, col, editValueParser)}
                     onBlur={this.onCreateCellBlur(rowIndex, col)}
-                    onFocus={this.onCellFocus}
+                    onFocus={this.onCellFocus('create', 'number', rowIndex, column.columnKey)}
                     onKeyDown={this.onCreateCellKeyDown}
                     inputRef={this.handleCreateCellRef(rowIndex, col)}
                     id={`ocDatagridCreateInput-${this.props.grid.id}-${column.columnKey}-${rowIndex}`}
@@ -713,6 +725,7 @@ export default class DataGrid extends React.PureComponent {
                     value={this.getEditItemValue(rowIndex, col)}
                     onChange={this.onEditCellValueChange(rowIndex, col, editValueParser)}
                     onBlur={this.onEditCellBlur(rowIndex, col, editValueParser)}
+                    onFocus={this.onCellFocus('edit', col.componentType, rowIndex, column.columnKey)}
                     onKeyDown={this.onEditCellKeyDown(col, rowIndex)}
                     inputRef={this.handleEditCellRef(rowIndex, col)}
                     id={`ocDatagridEditInput-${this.props.grid.id}-${column.columnKey}-${rowIndex}`}
@@ -773,6 +786,7 @@ export default class DataGrid extends React.PureComponent {
                     value={this.getEditItemValue(rowIndex, col)}
                     onChange={this.onEditCellValueChange(rowIndex, col, editValueParser)}
                     onBlur={this.onEditCellBlur(rowIndex, col)}
+                    onFocus={this.onCellFocus('edit', col.componentType, rowIndex, column.columnKey)}
                     searchable={selectOptions && (selectOptions.length > 9)}
                     clearable={!col.isRequired}
                     backspaceRemoves={false}
@@ -856,6 +870,7 @@ export default class DataGrid extends React.PureComponent {
                       id: `ocDatagridEditInput-${this.props.grid.id}-${column.columnKey}-${rowIndex}`,
                       onKeyDown: this.onEditCellKeyDown(col, rowIndex),
                       onBlur: this.onEditCellBlur(rowIndex, col),
+                      onFocus: this.onCellFocus('edit', col.componentType, rowIndex, column.columnKey),
                     }}
                     {...col.editComponentProps}
                     disabled={this.getComponentDisabledState(rowIndex, col, 'edit')}
@@ -914,6 +929,7 @@ export default class DataGrid extends React.PureComponent {
                     value={this.getEditItemValue(rowIndex, col)}
                     onChange={this.onEditCellValueChange(rowIndex, col, editValueParser)}
                     onBlur={this.onEditCellBlur(rowIndex, col)}
+                    onFocus={this.onCellFocus('edit', col.componentType, rowIndex, column.columnKey)}
                     searchable={false}
                     clearable={!col.isRequired}
                     backspaceRemoves={false}